use a sage.database.hdt package for hdt connector & iterator
